cleanup and create packages
