+++ exo-create/types/types.go
@@ -1,0 +1,6 @@
+package types
+
+// AppConfig represents the configuration for an application
+type AppConfig struct {
+	AppName, AppVersion, ExocomVersion, AppDescription string
+}
